firebase_function/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.*
flask==3.*
